# speed_up_/test-project/018-HAI/idp/internal/app/dto/user_dto.go
package dto

import (
	"github.com/google/uuid"
)

type UserDTO struct {
	ID       uuid.UUID `json:"id,omitempty"`
	Email    string    `json:"email"`
	Password string    `json:"password,omitempty"`
}