# speed_up_/test-project/018-HAI/idp/internal/app/dto/user_request.go
package dto

import "github.com/google/uuid"

type UserRequest struct {
	ID       uuid.UUID `json:"id,omitempty"`
	Email    string    `json:"email,omitempty"`
	Password string    `json:"password,omitempty"`
}